wip: create generator project
Creates a generator project that catches all references to generic names
and generates warnings. This is the first step in catching references to
the AddDefaultParser generic method and replacing them with calls to
add explicitly constructed parsers that use either the TryParse or
string constructors on the associated type. This avoids the need to use
reflection and makes the code safer for trimming, as well as allowing
us to raise an error if the type doesn't support a default parser.